+++ projects/visual-inference/student_detector/quad_geometry.py
@@ -430,7 +430,7 @@ def fit_quad_from_points(points: Tensor) -> Tensor:
     center = value.mean(dim=0)
     centered = value - center
     covariance = centered.transpose(0, 1) @ centered / max(value.shape[0] - 1, 1)
-    eigenvalues, eigenvectors = torch.linalg.eigh(covariance)
+    eigenvalues, eigenvectors = torch.linalg.eigh(covariance.to(torch.float32))
     major = eigenvectors[:, torch.argmax(eigenvalues)]
     major = major / torch.linalg.vector_norm(major).clamp(min=_EPS)
     minor = torch.stack((-major[1], major[0]))
